+++ packages/eslint-plugin/package.json
@@ -7,9 +7,19 @@
     "directory": "packages/eslint-plugin"
   },
   "license": "MIT",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "build": "echo 'Nothing to build..'"
+    "build": "yarn build:js && yarn build:types",
+    "build:js": "babel src -d dist --extensions \".js,.jsx,.ts,.tsx\"",
+    "build:types": "tsc --build --verbose",
+    "build:watch": "nodemon --watch src --ext \"js,jsx,ts,tsx\" --ignore dist --exec \"yarn build\"",
+    "prepublishOnly": "NODE_ENV=production yarn build",
+    "test": "jest src",
+    "test:watch": "yarn test --watch"
   },
   "devDependencies": {
     "@babel/cli": "7.21.5",
@@ -21,6 +31,9 @@
   },
   "gitHead": "3905ed045508b861b495f8d5630d76c7a157d8f1",
   "dependencies": {
+    "@babel/core": "7.22.1",
+    "@babel/runtime-corejs3": "7.22.3",
+    "core-js": "3.30.2",
     "eslint": "8.42.0"
   }
 }
