--- packages/eslint-plugin/src/index.js
@@ -1,7 +1,0 @@
-const processEnvComputed = require('./process-env-computed')
-
-module.exports = {
-  rules: {
-    'process-env-computed': processEnvComputed,
-  },
-}
