+++ packages/eslint-plugin/src/index.ts
@@ -1,0 +1,7 @@
+import processEnvComputed from './process-env-computed'
+
+export default {
+  rules: {
+    'process-env-computed': processEnvComputed,
+  },
+}
