+++ packages/eslint-plugin/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "extends": "../../tsconfig.compilerOption.json",
   "compilerOptions": {
+    "strict": true,
     "baseUrl": ".",
     "rootDir": "src",
     "tsBuildInfoFile": "dist/tsconfig.tsbuildinfo",
     "outDir": "dist",
   },
-  "include": ["src/**/*"],
+  "include": ["src"],
 }
